requirements.txt
apache-airflow[airbyte]
